+++ packages/components/composites/tests/styledSectionTest.js
