+++ packages/yoast-components/app/ButtonsWrapper.js
@@ -8,7 +8,7 @@ import { YoastLinkButton } from "../composites/Plugin/Shared/components/YoastLin
 import { BaseButton, Button, IconButton, IconsButton } from "../composites/Plugin/Shared/components/Button";
 import IconButtonToggle from "../composites/Plugin/Shared/components/IconButtonToggle";
 import { BaseLinkButton, LinkButton } from "../composites/Plugin/Shared/components/LinkButton";
-import FormButton from "../forms/Button";
+import FormButton from "../../components/Button";
 import IconLabelledButton from "../composites/Plugin/Shared/components/IconLabelledButton";
 
 const ButtonsContainer = styled.div`
